+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "gitpush"
-version = "0.1.8"
+version = "0.1.9"
 edition = "2021"
 authors = ["Waldir Borba Junior <wborbajr@gmail.com>"]
