+++ src/version.rs
@@ -1,3 +1,4 @@
+#[warn(dead_code)]
 pub fn show_version() -> String {
     const NAME: &str = env!("CARGO_PKG_NAME");
     const VERSION: &str = env!("CARGO_PKG_VERSION");
